requirements.txt
pandas
requests
python-dotenv
openai
neverbounce-sdk
langchain>=0.2.0
langchain-huggingface>=0.0.8
pydantic>=2.0.0
requests>=2.31.0
langchain-google-genai